Refactored most of rewriter implementation into base class
