+++ Traction/Rewriters/AttributeRewriters/NonEmptyRewriter.cs
@@ -1,6 +1,4 @@
-﻿using System;
-using System.Linq;
-using System.Text;
+﻿using System.Linq;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
@@ -14,45 +12,6 @@ namespace Traction {
 
         public NonEmptyRewriter(SemanticModel model, ICompileContext context)
             : base(model, context) { }
-
-        protected override StatementSyntax CreatePrecondition(TypeInfo type, string parameterName, Location location) {
-            if (parameterName == null) throw new ArgumentNullException(nameof(parameterName));
-            if (location == null) throw new ArgumentNullException(nameof(location));
-
-            var text = GetPreconditionText(parameterName, type.FullName());
-            var statement = SyntaxFactory.ParseStatement(text);
-
-            return SyntaxFactory.Block(statement);
-        }
-
-        protected override StatementSyntax CreatePostcondition(TypeInfo returnType, ReturnStatementSyntax node, Location location) {
-            if (node == null) throw new ArgumentNullException(nameof(node));
-            if (location == null) throw new ArgumentNullException(nameof(location));
-
-            var returnedExpression = node.ChildNodes().FirstOrDefault();
-            var tempVariableName = IdentifierFactory.CreatePostconditionLocal(node, model);
-            var text = GetPostconditionText(returnType.FullName(), returnedExpression.ToString(), tempVariableName);
-            var statement = SyntaxFactory.ParseStatement(text);
-            return statement;
-        }
-
-        private string GetPreconditionText(string parameterName, string parameterTypeName) {
-            var sb = new StringBuilder();
-            sb.AppendLine($"if ({GetConditionExpression(parameterName, null)})");
-            sb.AppendLine($"    throw new global::Traction.PreconditionException(\"{ExceptionMessage}\", nameof({parameterName}));");
-            return sb.ToString();
-        }
-
-        private string GetPostconditionText(string returnTypeName, string returnedExpression, string tempVarName) {
-            var sb = new StringBuilder();
-            sb.AppendLine("{");
-            sb.AppendLine($"    {returnTypeName} {tempVarName} = {returnedExpression};");
-            sb.AppendLine($"    if ({GetConditionExpression(tempVarName, null)})");
-            sb.AppendLine($"        throw new global::Traction.PostconditionException(\"{ExceptionMessage}\");");
-            sb.AppendLine($"    return {tempVarName};");
-            sb.AppendLine("}");
-            return sb.ToString();
-        }
 
         protected override string ExceptionMessage => "Sequence cannot be empty.";
