+++ Traction/Rewriters/AttributeRewriters/NonNullRewriter.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.Linq;
-using System.Text;
-using Microsoft.CodeAnalysis;
+﻿using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 
@@ -14,45 +11,7 @@ namespace Traction {
 
         public NonNullReWriter(SemanticModel model, ICompileContext context)
             : base(model, context) { }
-
-        protected override StatementSyntax CreatePrecondition(TypeInfo type, string parameterName, Location location) {
-            if (parameterName == null) throw new ArgumentNullException(nameof(parameterName));
-            if (location == null) throw new ArgumentNullException(nameof(location));
-
-            var text = GetPreconditionText(parameterName);
-            var statement = SyntaxFactory.ParseStatement(text);
-
-            return SyntaxFactory.Block(statement);
-        }
-
-        protected override StatementSyntax CreatePostcondition(TypeInfo returnType, ReturnStatementSyntax node, Location location) {
-            if (node == null) throw new ArgumentNullException(nameof(node));
-            if (location == null) throw new ArgumentNullException(nameof(location));
-
-            var returnedExpression = node.ChildNodes().FirstOrDefault();
-            var tempVariableName = IdentifierFactory.CreatePostconditionLocal(node, model);
-            var text = GetPostconditionText(returnType.FullName(), returnedExpression.ToString(), tempVariableName);
-            return SyntaxFactory.ParseStatement(text);
-        }
-
-        private string GetPreconditionText(string parameterName) {
-            var sb = new StringBuilder();
-            sb.AppendLine($"if ({GetConditionExpression(parameterName, null)})");
-            sb.AppendLine($"    throw new global::Traction.PreconditionException(\"{ExceptionMessage}\", nameof({parameterName}));");
-            return sb.ToString();
-        }
-
-        private string GetPostconditionText(string returnTypeName, string returnedExpression, string tempVarName) {
-            var sb = new StringBuilder();
-            sb.AppendLine("{");
-            sb.AppendLine($"    {returnTypeName} {tempVarName} = {returnedExpression};");
-            sb.AppendLine($"    if ({GetConditionExpression(tempVarName, null)})");
-            sb.AppendLine($"        throw new global::Traction.PostconditionException(\"{ExceptionMessage}\");");
-            sb.AppendLine($"    return {tempVarName};");
-            sb.AppendLine("}");
-            return sb.ToString();
-        }
-
+        
         protected override string ExceptionMessage => "Value cannot be null.";
 
         protected override ExpressionSyntax GetConditionExpression(string expression, string expressionType) {
