+++ Traction/Rewriters/ContractRewriterBase.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
+using System.Text;
 
 namespace Traction {
 
@@ -141,13 +142,41 @@ namespace Traction {
             }
         }
 
+        private StatementSyntax CreatePrecondition(TypeInfo parameterType, string parameterName, Location location) {
+            var text = GetPreconditionText(parameterName, parameterType.FullName());
+            var statement = SyntaxFactory.ParseStatement(text);
+            return SyntaxFactory.Block(statement);
+        }
+
+        private StatementSyntax CreatePostcondition(TypeInfo returnType, ReturnStatementSyntax node, Location location) {           
+            var returnedExpression = node.ChildNodes().FirstOrDefault();
+            var tempVariableName = IdentifierFactory.CreatePostconditionLocal(node, model);
+            var text = GetPostconditionText(returnType.FullName(), returnedExpression.ToString(), tempVariableName);
+            var statement = SyntaxFactory.ParseStatement(text);
+            return statement;
+        }
+
+        private string GetPreconditionText(string parameterName, string parameterTypeName) {
+            var sb = new StringBuilder();
+            sb.AppendLine($"if ({GetConditionExpression(parameterName, parameterTypeName)})");
+            sb.AppendLine($"    throw new global::Traction.PreconditionException(\"{ExceptionMessage}\", nameof({parameterName}));");
+            return sb.ToString();
+        }
+
+        private string GetPostconditionText(string returnTypeName, string returnedExpression, string tempVarName) {
+            var sb = new StringBuilder();
+            sb.AppendLine("{");
+            sb.AppendLine($"    {returnTypeName} {tempVarName} = {returnedExpression};");
+            sb.AppendLine($"    if ({GetConditionExpression(tempVarName, returnTypeName)})");
+            sb.AppendLine($"        throw new global::Traction.PostconditionException(\"{ExceptionMessage}\");");
+            sb.AppendLine($"    return {tempVarName};");
+            sb.AppendLine("}");
+            return sb.ToString();
+        }
+        
         protected abstract ExpressionSyntax GetConditionExpression(string expression, string expressionType);
 
         protected abstract string ExceptionMessage { get; }
-
-        protected abstract StatementSyntax CreatePrecondition(TypeInfo parameterType, string identifier, Location location);
-
-        protected abstract StatementSyntax CreatePostcondition(TypeInfo returnType, ReturnStatementSyntax node, Location location);
 
         protected abstract bool IsValidType(TypeInfo type);
